+++ packages/vite-plugin-branch/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @osdk/vite-plugin-branch
 
+## 2.71.0
+
 ## 2.70.0
 
 ## 2.69.0